backend working - delete CategoryConfig - modify test
