+++ donetop-main/src/test/java/com/donetop/main/api/common/DraftBase.java
@@ -2,18 +2,11 @@ package com.donetop.main.api.common;
 
 import com.donetop.domain.entity.draft.Draft;
 import com.donetop.domain.entity.folder.DraftFolder;
-import com.donetop.domain.entity.user.User;
 import com.donetop.enums.folder.FolderType;
-import com.donetop.enums.user.RoleType;
 import com.donetop.common.service.storage.LocalFileUtil;
 import com.donetop.common.service.storage.Resource;
 import com.donetop.common.service.storage.StorageService;
 import com.donetop.repository.draft.DraftRepository;
-import com.donetop.repository.user.UserRepository;
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.response.Response;
-import org.json.JSONObject;
 import org.junit.jupiter.api.AfterAll;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.FileSystemUtils;
@@ -24,9 +17,7 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.donetop.main.api.form.FormAPIController.URI.LOGIN;
-
-public class DraftBase extends IntegrationBase {
+public class DraftBase extends UserBase {
 
 	@Autowired
 	protected DraftRepository draftRepository;
@@ -34,13 +25,9 @@ public class DraftBase extends IntegrationBase {
 	@Autowired
 	protected StorageService<DraftFolder> draftFolderStorageService;
 
-	@Autowired
-	protected UserRepository userRepository;
-
 	@AfterAll
-	void afterAll() throws IOException {
+	void clearDraftBase() throws IOException {
 		draftRepository.deleteAll();
-		userRepository.deleteAll();
 		FileSystemUtils.deleteRecursively(Path.of(testStorage.getRoot()));
 	}
 
@@ -107,24 +94,6 @@ public class DraftBase extends IntegrationBase {
 		for (final FolderType folderType : folderTypes)
 			draftFolderStorageService.saveOrReplace(resources, draftFolderStorageService.addNewFolderOrGet(draft, folderType));
 		return draftRepository.save(draft);
-	}
-
-	protected Response doLoginWith(final User user) throws Exception {
-		final JSONObject loginBody = new JSONObject();
-		loginBody.put("username", user.getEmail());
-		loginBody.put("password", user.getPassword());
-		return RestAssured.given(this.spec).when()
-			.contentType(ContentType.JSON)
-			.body(loginBody.toString())
-			.post(LOGIN);
-	}
-
-	protected User saveUser(final String name, final RoleType roleType) {
-		User normal = User.builder()
-			.email(name + "@test.com")
-			.name(name)
-			.password("password").build().updateRoleType(roleType);
-		return userRepository.save(normal);
 	}
 
 }
